Conflicts:
	nakeduml-core/org.nakeduml.generators/velocity/net/sf/nakeduml/javageneration/hibernate/HibernateConfigGenerator.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/AbstractJavaTransformationStep.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/JavaTransformationPhase.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/JavaTransformationStep.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/basicjava/Java5ModelGenerator.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/basicjava/PersistentNameMapGenerator.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/hibernate/HibernateAnnotator.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/jbpm5/AbstractEventHandlerInserter.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/jbpm5/Jbpm5Util.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/jbpm5/ProcessStepEnumerationImplementor.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/jbpm5/activity/ActivityNodeEnumerationImplementor.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/jbpm5/activity/ActivityProcessImplementor.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/oclexpressions/PreAndPostConditionGenerator.java
	nakeduml-core/org.nakeduml.java.generator/javageneration/net/sf/nakeduml/javageneration/persistence/JpaAnnotator.java
	nakeduml-core/org.nakeduml.metamodels/feature/net/sf/nakeduml/feature/NakedUmlConfig.java
	nakeduml-core/org.nakeduml.metamodels/lib/nakeduml-runtime-domain-1.0.0.5-SNAPSHOT.jar
	nakeduml-core/org.nakeduml.metamodels/linkage/net/sf/nakeduml/linkage/NakedParsedOclStringResolver.java
	nakeduml-core/org.nakeduml.metamodels/metamodel/net/sf/nakeduml/metamodel/core/IParameterOwner.java
	nakeduml-core/org.nakeduml.tinker.generator/javageneration/org/nakeduml/tinker/composition/CopyMethodImplementor.java
	nakeduml-core/org.nakeduml.tinker.generator/javageneration/org/nakeduml/tinker/composition/tinker/TinkerCompositionNodeStrategy.java
	nakeduml-runtime/nakeduml-runtime-adaptor/.classpath
	nakeduml-runtime/nakeduml-runtime-adaptor/pom.xml
	nakeduml-runtime/nakeduml-runtime-adaptor/src/main/java/org/nakeduml/audit/AuditCapturer.java
	nakeduml-runtime/nakeduml-runtime-adaptor/src/main/java/org/nakeduml/environment/adaptor/AbstractSignalMdb.java
	nakeduml-runtime/nakeduml-runtime-adaptor/src/main/java/org/nakeduml/environment/adaptor/EntitySignalMdb.java
	nakeduml-runtime/nakeduml-runtime-adaptor/src/main/java/org/nakeduml/environment/adaptor/SignalDispatcher.java
	nakeduml-runtime/nakeduml-runtime-domain/pom.xml
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/environment/ISignalDispatcher.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/environment/PersistentNameClassMap.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/environment/SignalToDispatch.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/event/AbstractNakedUmlEvent.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/hibernate/domain/EventDispatcher.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/hibernate/domain/FlushEntityListener.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/runtime/domain/AbstractEntity.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/runtime/domain/AbstractEnum.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/runtime/domain/AbstractProcessStep.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/runtime/domain/AbstractSignal.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/workinghours/TimeUnit.java
	nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/workinghours/WorkDayType.java
	nakeduml-runtime/nakeduml-test-adaptor/.classpath
	nakeduml-runtime/nakeduml-test-adaptor/pom.xml
	nakeduml-runtime/pom.xml

+++ nakeduml-runtime/nakeduml-runtime-tinker-domain/src/main/java/org/nakeduml/runtime/domain/TinkerAuditNode.java
@@ -2,8 +2,7 @@ package org.nakeduml.runtime.domain;
 
 
 public interface TinkerAuditNode extends TinkerNode {
-	int getTransactionNo();
-//	List<? extends TinkerAuditNode> getNextAuditEntries();
+	Long getTransactionNo();
 	TinkerAuditNode getNextAuditEntry();
 	TinkerAuditableNode getOriginal();
 }
